Simplify TalonAction and rename file
